requirements.txt
infi.systray==0.1.12
python-dotenv==1.0.1
aiogram==3.3.0
SQLAlchemy~=2.0.25